+++ .lintstagedrc.cjs
@@ -1,4 +1,4 @@
 /* eslint-disable */
-const configs = require("./dist/index");
+const configs = require("./dist-cjs/index.js");
 
 module.exports = configs.lintStaged;